+++ demo/WalkingTec.Mvvm.Demo/DataContext.cs
@@ -70,7 +70,7 @@ namespace WalkingTec.Mvvm.Demo
                 var user = new FrameworkUser
                 {
                     ITCode = "admin",
-                    Password = Utils.GetMD5String("000000"),
+                    Password = PasswordHashHelper.HashPassword("000000"),
                     IsValid = true,
                     Name = "Admin",
                     TenantCode = TenantCode