+++ demo/WalkingTec.Mvvm.ReactDemo/DataContext.cs
@@ -46,7 +46,7 @@ namespace WalkingTec.Mvvm.ReactDemo
                 var user = new FrameworkUser
                 {
                     ITCode = "admin",
-                    Password = Utils.GetMD5String("000000"),
+                    Password = PasswordHashHelper.HashPassword("000000"),
                     IsValid = true,
                     Name = "Admin"
                 };