+++ demo/WalkingTec.Mvvm.Vue3Demo/DataContext.cs
@@ -50,7 +50,7 @@ namespace WalkingTec.Mvvm.Vue3Demo
                 var user = new FrameworkUser
                 {
                     ITCode = "admin",
-                    Password = Utils.GetMD5String("000000"),
+                    Password = PasswordHashHelper.HashPassword("000000"),
                     IsValid = true,
                     Name = "Admin"
                 };